requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# LangGraph and LangChain ecosystem - Updated with compatible versions
langgraph>=0.3.27
langchain-core>=0.2.38
langchain-openai>=0.2.8
openai>=1.68.0
python-dotenv>=1.0.0

# Additional LangGraph ecosystem packages (optional but recommended)
langsmith>=0.1.63
langchain-community>=0.2.0
